Added name parser and formatter to the standard distribution. Build: 20130326-build-1.3-005
Added "sync" option to any class that takes an onLoad parameter. That
way, it is explicit whether or not you want to do asynchronous or
synchronous loading of the missing locale data.
Changed util's getLocFiles so that the prefix is no longer included.
This way, the dynamic loader function can add its own prefix and load
the file from whatever directory it is stored in.
Fixed errors in the json syntax of some of the timezone.jf files.
Get named resources to load correctly in a dynamic load environment.