Added work around hack to make loading files synchronous on device.
This should be taken out again when enyo.Ajax is fixed. Also added in
ability to return pseudo-ized strings for non-English locales when a
string is missing. This will help QA tell if a string is (1) translated
(2) $L'ed but not yet translated, and (3) not $L'ed. Previously, they
could not tell the difference between cases 2 and 3.