GF-12831: TypeError while loading strings.
Avoid using undefined objects to load strings inside of the
$L function. In some circumstances, the resource bundle is not
ready yet, and the call to $L.rb.getString will fail with a
TypeError because $L.rb is not defined yet. This puts in some
defensive code to guard against that. Instead, the string
parameter (in English probably) will be returned as-is.