GF-3297: load strings.json from a different path
The glue code now looks at the loadParams.root option which
tells it which path to look in for the localized files. If
this option is not specified, it defaults to the previous
path of "resources" underneath the dir where the index.html
is located.

Enyo-DCO-1.1-Signed-Off-By: Edwin Hoogerbeets (edwin.hoogerbeets@lge.com)