Update to ilib build 20131025-build-4.0-001
This build contains code for the Thai solar calendar, which is
used in Thailand as the official calendar. This has the same
structure, months and such as the Gregorian calendar, but the
year 0 in the Thai solar calendar is the year Buddha died rather
than the year that Jesus was born. (This is 543 years earlier.)