Build 20131121-build-4.0-004 of ilib
This fixes GF-45592: 'root' parameter in loadParams

When code attempts to load resources from a different directory, it
needs to give the root parameter inside of the loadParams to tell
the loading code what is the root directory where it is supposed to
find all the localizations of the requested file. However, this was
not working correctly if you already loaded the strings for your
own app because the caching code only keyed off the locale and got
a cache hit. Now the caching code keys off a combination of the
locale, the name of the file being loaded, and a hash of the
loadParams object. If any of those differ, then the code will load
the new file instead of getting an incorrect cache hit.