Update to 20131204-build-4.0-005
Update to ilib build 20131204-build-4.0-005
Add a copy constructor to ilib.Locale to fix a problem when there
are no translation resources available.
Fixes GF-53464: update to Korean date formatting
Make sure that ilib.loadData caches locale data based on a key that
depends on the locale + file name + hash of load parameters, so that
if any of those change, there is a cache miss.