Update to 20140202-build-5.0-004
This build fixes GF-55464 (durations cut off because they contain the
AM/PM text) for all locales, including Bengali where it was originally
reported.

Other changes include:

- fixed 12-hour time formats for various locales that were using the
24-hour template characters "H" and "HH" instead of "h" and "hh".

- fixed completely broken time formats for Marathi (mr) and Thai (th)

- If a locale uses native digits by default for formatting numbers,
ilib.LocaleInfo.getNativeDigits() was not returning those native digits.

- In the date formatter, if the option "useNative: false" was given,
but the locale uses native digits by default, it would format the date
with native digits anyways. Now, if the useNative option is given,
it is honoured even if the locale normally uses native digits.

- Added ilib.LocaleInfo.getDigitsStyle() to return whether the locale
uses western digits or native digits. If the locale has native digits,
but they are not commonly used currently, this method can return that
information as well.

- Added useNative option to the duration formatter constructor to
explicitly say whether or not to use native digits. If this option is
left out, the formatter will default to the style of digits specified
by the locale.