Update to ilib build 20140515-build-6.0-002
The GregDate object was not calculating the offset when the time zone is "local" and the instance was created with something other than the time components (ie. it was created with a julian day or unix time for example). The methods that used the offset, such as getDayOfWeek(), were therefore only calculating things in UTC.

Added various unit tests to verify that it is working properly now.

Update the methods on ilib.Date that set the date components so that they recalculate the RD based on the new component.

The date formatter no longer throws an exception if you attempt to format a date object that is in a different calendar than the formatter. Instead, it just converts it to the formatter's calendar and formats the result.

Made the demo site work properly with time zones.

Fixed problem where zone info was being loaded dynamically per locale, and another problem when constructing a time zone instance with an id that is specified with an ilib.String instead of a javascript string.

Unit tests added to verify the new functionality, and all tests pass.