Update to 20140520-build-6.0-003
This is a simple change to add the ability to the time zone class to
specify the DST transition times using a julian day rather than via
the transition rules. This allows us to use the system zone info which
do not contain the rules, but pre-calculated dates when the
transitions happen.