Update to ilib build 20140530-build-6.0-004
This fixes the following things:

* Duration formats in Arabic for the Gregorian calendar are fixed to
use full-length names for the components instead of abbreviations
which are never used in Arabic
* Duration formats now use the proper names of components depending
on the number
* Date and date range formats are fixed to appear in the correct
order now. (dd/MM/yyyy instead of yyyy/MM/dd)
* JSDoc comments are updated, and the JSDoc extractor updated to be
able to deal with the Google closure variable type names. This means
the JS docs are much more usable now