Fixed resource loading in the glue code
Fixed resource loading in the glue code so that it can deal with
root directories that do not contain an ilibmanifest.json file.
Also, make sure when we are updating the locale when loading in
an app into the container, we invalidate the manifest cache to
force it to load the manifests again.

Conflicts:
	glue.js