Update RTL date & date range formatting
Fixed the date and date range formatting for RTL languages by
introducing the U+200F (RTL) character at the beginning of
formats that start with an LTR character. This way, webkit does
not misinterpret the direction to start writing the text.

Also updated the numeric-based Farsi date range formats (mostly
these are the short and medium formats) to show on the screen
with the start on the left and end on the right. The textual
formats were left alone, as they are normally read from right
to left.