Update to ilib build 20141029-build-9.0-001
This adds in a few more locales for the phone number parsing/formatting,
as well as a few bug fixes. It also adds in the unit conversion routines
(eg. Celcius -> Fahrenheit, or feet -> meters), as well as the
units formatter with translations. (eg. measurement of 500 meters is
auto-converted and formatted for locale en-US as "1,640.42 feet")
Also fixed the performance of the CType routines which were causing a
slow down with many other classes that used CType to do parsing.