Update to ilib 20141218-build-9.0-006
Fix the order of the translations of the days of the week for mn_Cyrl_MN and ku-Arab-IQ. Add Korean phonenumber formatting unit test cases with google pause and wait characters for extensions. Add more phone number normalization unit tests to test correctness for Korea, Japan, Brazil, Russia and Taiwan. Added test to avoid an exception in the glyphstring constructor Add pseudomap/pseudo-locale related changes so that we can use any locale as our psuedo locale on a temporary basis.