Changes to get dynamic code loading working
Changes to get the dynamic code loading working properly
with the enyo sampler app. The package.json file can load
in any size of ilib it wants from ilibglobal.js (almost nothing)
to core to standard to full, and various amounts of ilib
will be preloaded. The portion of ilib that is not preloaded
will still be available for dynamic loading.